Moved bitboard struct into its own header file, fixed requested changes on PR
